Fixed bugs, XML validated, WordDocList saveable
I fixed some errors that were preventing the XML output to be well formed. I edited the schema so that the XML document would be valid against it, at the bottom of the XMLReturnsTestScripts.py I provided a way to save the XML code from the a WordDocList structure as an XML document. A WordDocList is simply a list of objects representing the full word document. The WordDocList should be created by the Parser.