Procedure more flexible, ProcedureParser
I modified the Procedure class to contain elements, so it can now contain TextualElements, GraphicalElements, and Tables, all in the same list, elementList. Added corresponding setters and getters and removed the setters and getters for the steps and graphics in Procedure. Incomplete ProcedureParser, but I have made progress. Currently able to identify procedures, but the extraction method doesn't work as intended.