I edited Luis's Parser to be functional
Luis's parser had several bugs. I fixed the ones that I encountered. I reorganized it as a class. I added a paragraphParse() method that parses all the paragraphs in the document and returns them as a list of TextualElements(). I used several of the methods Luis created, but (as previously mentioned) I had to fix some bugs.